displayPosts called
calling displayPosts or displayUserPosts after delete, comment edit etc to make the DOM update
