+++ js/api/posts/create.mjs
@@ -1,6 +1,7 @@
 import { API_SOCIAL_URL } from "../api_constants.mjs";
 import { authFetch } from "../authFetch.mjs";
 import { showMessage } from "../../utils/messages.mjs";
+import { getPosts } from "./get.mjs";
 
 const action = "/posts";
 const method = "POST";
@@ -25,8 +26,8 @@ export async function createPost(postData) {
       throw new Error(`Failed to create post: ${response.status}`);
     }
 
-    const post = await response.json();
-    return post;
+    const updatedPosts = await getPosts(); // Fetch the updated list of posts
+    return updatedPosts; // Return the updated list of posts
   } catch (error) {
     console.error("Error creating post:", error);
     showMessage("Failed to create post. Please try again later.", "error");
