+++ js/handlers/index.mjs
@@ -12,12 +12,10 @@ export * from "./observers/reactions.mjs";
 export * from "./posts/displayPosts.mjs";
 export * from "./posts/createPost.mjs";
 export * from "./posts/updatePost.mjs";
-export * from "./posts/editButton.mjs";
-export * from "./posts/modalEditPost.mjs";
 export * from "./posts/delete.mjs";
 export * from "./posts/comment.mjs";
 export * from "./posts/likeClick.mjs";
 export * from "./posts/filters.mjs";
 export * from "./posts/search.mjs";
 export * from "./posts/displaySinglePost.mjs";
-export * from "./posts/commentDelete.mjs";
+export * from "./posts/commentDelete.mjs"
