+++ js/handlers/posts/delete.mjs
@@ -2,10 +2,7 @@ import { getPostById } from "../../api/posts/get.mjs";
 import { removePost } from "../../api/posts/delete.mjs";
 import { showMessage } from "../../utils/messages.mjs";
 import { createSinglePostElement } from "../../templates/index.mjs";
-import {
-  restoreScrollPosition,
-  storeScrollPosition,
-} from "../../utils/scrollPosition.mjs";
+import { displayPosts, displayUserPosts } from "../index.mjs";
 
 export async function handleDeleteButtonClick(event, postId) {
   event.preventDefault();
@@ -23,22 +20,24 @@ export async function handleDeleteButtonClick(event, postId) {
 
     const deletePostElement = createSinglePostElement(postData);
 
-   
     modalContent.appendChild(deletePostElement);
     // Event listener to the confirm delete button
     const confirmDeleteButton = document.getElementById("confirmDeleteButton");
     confirmDeleteButton.addEventListener("click", async () => {
       try {
-        storeScrollPosition();
         // Call the removePost function to delete the post
         await removePost(postId);
         deletePostModal.hide();
 
         showMessage("Post deleted successfully!", "success");
-        setTimeout(() => {
-          window.location.reload();
-        }, 2000);
-        restoreScrollPosition();
+   
+        
+        if (document.querySelector("#userPosts")) {
+          displayUserPosts();
+        } else {
+          displayPosts();
+        }
+        
       } catch (error) {
         const errorMessage = "Could not delete post:" + error.message;
